refactor: 불필요한 테스트 제거 및 패키지 정렬
