+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -1,4 +1,4 @@
-package chess.piece;
+package chess.domain.piece;
 
 import chess.domain.Team;
 import chess.domain.piece.Pawn;
