+++ src/test/java/chess/domain/position/PositionTest.java
@@ -1,4 +1,4 @@
-package chess.position;
+package chess.domain.position;
 
 import chess.domain.piece.Direction;
 import chess.domain.position.Position;
