+++ src/test/java/chess/domain/position/RowTest.java
@@ -1,4 +1,4 @@
-package chess.position;
+package chess.domain.position;
 
 import chess.domain.position.Row;
 import org.junit.jupiter.api.DisplayName;
